normalize component names; remove dead code
