+++ client-admin/src/components/conversation-admin/ShareAndEmbed.js
@@ -1,13 +1,13 @@
 // Copyright (C) 2012-present, The Authors. This program is free software: you can redistribute it and/or  modify it under the terms of the GNU Affero General Public License, version 3, as published by the Free Software Foundation. This program is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU Affero General Public License for more details. You should have received a copy of the GNU Affero General Public License along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-import ConversationHasCommentsCheck from './conversation-has-comments-check'
+import ConversationHasCommentsCheck from './ConversationHasCommentsCheck'
 import Url from '../../util/url'
 import { useSelector } from 'react-redux'
 import { Link, useParams } from 'react-router'
 import { Heading, Text, Box } from 'theme-ui'
 import ComponentHelpers from '../../util/component-helpers'
-import NoPermission from './no-permission'
-import ParticipantXids from './participant-xids'
+import NoPermission from './NoPermission'
+import ParticipantXids from './ParticipantXids'
 
 const ShareAndEmbed = () => {
   const params = useParams()
