--- client-admin/src/components/conversation-admin/data-export.js
@@ -1,136 +1,0 @@
-// Copyright (C) 2012-present, The Authors. This program is free software: you can redistribute it and/or  modify it under the terms of the GNU Affero General Public License, version 3, as published by the Free Software Foundation. This program is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU Affero General Public License for more details. You should have received a copy of the GNU Affero General Public License along with this program.  If not, see <http://www.gnu.org/licenses/>.
-
-import { useState, useRef } from 'react'
-import { useSelector, useDispatch } from 'react-redux'
-import { startDataExport } from '../../actions'
-import dateSetupUtil from '../../util/data-export-date-setup'
-import { Heading } from 'theme-ui'
-import ComponentHelpers from '../../util/component-helpers'
-import NoPermission from './no-permission'
-
-const DataExport = () => {
-  const dispatch = useDispatch()
-  const zid_metadata = useSelector((state) => state.zid_metadata)
-
-  const times = dateSetupUtil()
-  const [state, setState] = useState({
-    ...times,
-    showHelpMessage: false,
-    untilEnabled: false
-  })
-
-  const exportSelectYearRef = useRef(null)
-  const exportSelectMonthRef = useRef(null)
-  const exportSelectDayRef = useRef(null)
-  const exportSelectHourRef = useRef(null)
-
-  const handleExportClicked = () => {
-    setState((prev) => ({ ...prev, showHelpMessage: true }))
-
-    const year = exportSelectYearRef.current.value
-    const month = exportSelectMonthRef.current.value
-    const dayOfMonth = exportSelectDayRef.current.value
-    const tz = exportSelectHourRef.current.value
-    const format = 'csv'
-
-    const dateString = [year, month, dayOfMonth, tz].join(' ')
-    const dddate = new Date(dateString)
-
-    dispatch(
-      startDataExport(
-        zid_metadata.zid_metadata.conversation_id,
-        format,
-        (dddate / 1000) << 0,
-        !!state.untilEnabled
-      )
-    )
-  }
-
-  const showHelpMessage = () => {
-    return (
-      <div>
-        <p>
-          Data from this conversation will be sent to your email. (This can take a little while,
-          especially for larger conversations).
-        </p>
-      </div>
-    )
-  }
-
-  const handleUntilToggled = () => {
-    setState((prev) => ({ ...prev, untilEnabled: !prev.untilEnabled }))
-  }
-
-  if (
-    ComponentHelpers.shouldShowPermissionsError({
-      zid_metadata: zid_metadata.zid_metadata,
-      loading: zid_metadata.loading
-    })
-  ) {
-    return <NoPermission />
-  }
-
-  return (
-    <div>
-      <Heading
-        as="h3"
-        sx={{
-          fontSize: [3, null, 4],
-          lineHeight: 'body',
-          mb: [3, null, 4]
-        }}>
-        Export
-      </Heading>
-      <div>
-        <p> Until: </p>
-        <input onClick={handleUntilToggled} type="checkbox" />
-        <select disabled={state.untilEnabled ? '' : 'disabled'} ref={exportSelectYearRef}>
-          {state.years.map((year, i) => {
-            return (
-              <option selected={year.selected} key={i} value={year.name}>
-                {year.name}
-              </option>
-            )
-          })}
-        </select>
-        <select disabled={state.untilEnabled ? '' : 'disabled'} ref={exportSelectMonthRef}>
-          {state.months.map((month, i) => {
-            return (
-              <option selected={month.selected} key={i} value={month.name}>
-                {month.name}
-              </option>
-            )
-          })}
-        </select>
-        <select disabled={state.untilEnabled ? '' : 'disabled'} ref={exportSelectDayRef}>
-          {state.days.map((day, i) => {
-            return (
-              <option selected={day.selected} key={i} value={day.name}>
-                {' '}
-                {day.name}{' '}
-              </option>
-            )
-          })}
-        </select>
-        <select disabled={state.untilEnabled ? '' : 'disabled'} ref={exportSelectHourRef}>
-          {state.tzs.map((tz, i) => {
-            return (
-              <option selected={tz.selected} key={i} value={tz.name}>
-                {' '}
-                {tz.name}{' '}
-              </option>
-            )
-          })}
-        </select>
-        <p>
-          By default, the entire dataset is returned. To limit the last timestamp returned, enter a
-          date here.
-        </p>
-        <button onClick={handleExportClicked}>Export</button>
-        {state.showHelpMessage ? showHelpMessage() : null}
-      </div>
-    </div>
-  )
-}
-
-export default DataExport
