+++ client-admin/src/components/conversation-admin/index.js
@@ -7,12 +7,12 @@ import { populateZidMetadataStore, resetMetadataStore } from '../../actions'
 import { Routes, Route, Link, useParams, useLocation } from 'react-router'
 import { useAuth } from 'react-oidc-context'
 
-import ConversationConfig from './conversation-config'
+import ConversationConfig from './ConversationConfig'
 import ConversationStats from './stats'
 import ModerateComments from './comment-moderation/'
 import TopicModeration from './topic-moderation/'
-import ShareAndEmbed from './share-and-embed'
-import Reports from './report/reports'
+import ShareAndEmbed from './ShareAndEmbed'
+import Reports from './report/Reports'
 import InviteTree from './InviteTree'
 import InviteCodes from './InviteCodes'
