+++ client-admin/src/components/conversation-admin/report/ReportsList.test.js
@@ -4,7 +4,7 @@ import { configureStore } from '@reduxjs/toolkit'
 import { ThemeUIProvider } from 'theme-ui'
 import { BrowserRouter as Router } from 'react-router'
 import theme from '../../../theme'
-import ReportsList from './reports-list'
+import ReportsList from './ReportsList'
 import PolisNet from '../../../util/net'
 import * as actions from '../../../actions'
 import { mockAuth } from '../../../test-utils'
