+++ client-admin/src/components/conversation-admin/stats/index.js
@@ -5,12 +5,12 @@ import { useState, useEffect, useRef } from 'react'
 import { useSelector, useDispatch } from 'react-redux'
 import { populateConversationStatsStore, populateZidMetadataStore } from '../../../actions'
 import { useAuth } from 'react-oidc-context'
-import NumberCards from './conversation-stats-number-cards'
-import Voters from './voters'
-import Commenters from './commenters'
+import NumberCards from './NumberCards'
+import Voters from './Voters'
+import Commenters from './Commenters'
 import { Heading, Box } from 'theme-ui'
 import ComponentHelpers from '../../../util/component-helpers'
-import NoPermission from '../no-permission'
+import NoPermission from '../NoPermission'
 import { useParams } from 'react-router'
 
 const ConversationStats = () => {
