+++ client-admin/src/components/conversation-admin/topic-moderation/index.js
@@ -2,16 +2,14 @@
 // Copyright (C) 2012-present, The Authors. This program is free software: you can redistribute it and/or  modify it under the terms of the GNU Affero General Public License, version 3, as published by the Free Software Foundation. This program is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU Affero General Public License for more details. You should have received a copy of the GNU Affero General Public License along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
 import ComponentHelpers from '../../../util/component-helpers'
-import NoPermission from '../no-permission'
+import NoPermission from '../NoPermission'
 import React from 'react'
 import { connect } from 'react-redux'
 import { Heading, Flex, Box } from 'theme-ui'
 import { Routes, Route, Link } from 'react-router-dom'
 
-import TopicTree from './topic-tree'
-import TopicDetail from './topic-detail'
-import TopicStats from './topic-stats'
-import ProximityVisualization from './proximity-visualization'
+import TopicTree from './TopicTree'
+import TopicStats from './TopicStats'
 
 const mapStateToProps = (state, ownProps) => {
   return {
