+++ client-admin/src/components/framework/logomark.test.js
@@ -1,7 +1,7 @@
 import { render } from '@testing-library/react'
 import { ThemeUIProvider } from 'theme-ui'
 import theme from '../../theme'
-import Logomark from './logomark'
+import Logomark from './Logomark'
 
 // Wrapper to provide theme context
 const renderWithTheme = (component) => {
