+++ client-admin/src/components/framework/spinner.test.js
@@ -1,5 +1,5 @@
 import { render } from '@testing-library/react'
-import Spinner from './spinner'
+import Spinner from './Spinner'
 
 describe('Spinner', () => {
   it('renders without crashing', () => {
